Final Commit
I set up all the font families at the beginning and the charset.  i changed the spacing for a lot of different things in the site. I also used several different font families and background colors. I changed the size and weight of many things as we I changed the anchor tags in navigation and all of its states.